Vredno smo radili

+++ routes/MealRoute.js
@@ -31,8 +31,11 @@ router.get('/:mealId', async (req, res) => {
 });
 
 router.post('/', async (req, res) => {
-    const {name, price, tag} = req.body;
+    if(!req.logged && req.user != 'admin'){
+        return res.status(403).json(getResponse(null, 'Unauthorized'));
+    }
 
+    const {name, price, tag} = req.body;
     // input validation - should create func for this
     if(name == '' || tag == ''){
         return res.status(400).json(getResponse(null, 'Bad Request'));
@@ -45,9 +48,13 @@ router.post('/', async (req, res) => {
             tag: tag
         });
         // save to db
-        savedMeal = await meal.save();
+        const savedMeal = await meal.save();
+        if(savedMeal){
+            return res.status(200).json(getResponse(savedMeal._doc, 'Success'));
+        } else {
+            return res.status(500).json(getResponse(null, 'Error while saving meal to db'));
+        }
         // send result
-        return res.status(200).json(getResponse(savedMeal._doc, 'Success'));
     } catch(err){
         console.log(err);
         return res.status(400).json(getResponse(null, err));
@@ -56,27 +63,10 @@ router.post('/', async (req, res) => {
 
 // edit
 router.put('/:mealId', async (req, res) => {
-    const {name, price, tag} = req.body;
-    if(name == '' || tag == ''){
-        return res.status(400).json(getResponse(null, 'Bad Request'));
+    if(!req.logged && req.user != 'admin'){
+        return res.status(403).json(getResponse(null, 'Unauthorized'));
     }
-    
-    try{
-        const id = req.params.mealId;
-        const meal = await Meal.findOneAndUpdate({_id: id}, { ...req.body }, {useFindAndModify: false});
-        if(meal){
-            return res.status(200).json(getResponse({ ...meal._doc, ...req.body }, 'Success'));
-        } else {
-            return res.status(404).json(getResponse(null, 'Not Found'));
-        }
-    } catch(err){
-        console.log(err);
-        return res.status(400).json(getResponse(null, err));
-    }
-});
 
-// edit
-router.patch('/:mealId', async (req, res) => {
     const {name, price, tag} = req.body;
     if(name == '' || tag == ''){
         return res.status(400).json(getResponse(null, 'Bad Request'));
@@ -98,10 +88,14 @@ router.patch('/:mealId', async (req, res) => {
 
 // delete
 router.delete('/:mealId', async (req, res) => {
+    if(!req.logged && req.user != 'admin'){
+        return res.status(403).json(getResponse(null, 'Unauthorized'));
+    }
+
     try{
         const deletedMeal = await Meal.findByIdAndDelete(req.params.mealId);
         if(deletedMeal){
-            return res.status(200).json(getResponse(deletedMeal._doc, 'Success'));
+            return res.status(200).json(getResponse(null, 'Success'));
         } else {
             return res.status(404).json(getResponse(null, 'Not Found'));
         }
